split threaded pms
